split DayContent into multiple files
